uploads: Support non-AWS S3-compatible server.
Boto3 does not allow setting the endpoint url from
the config file. Thus we create a django setting
variable (`S3_ENDPOINT_URL`) which is passed to
service clients and resources of `boto3.Session`.

We also set the `AWS_CONFIG_FILE` environmental
variable instead of the `BOTO_CONFIG` variable.
This is done because for looking up the configuration
values boto3 searches for the `~/.aws/config` file,
as mentioned in the docs.

Fixes #16246.